Extract First Party Hosts types to RUM
