+++ packages/core/src/index.tsx
@@ -8,8 +8,7 @@ import {
     DatadogProviderConfiguration,
     DdSdkReactNativeConfiguration,
     InitializationMode,
-    VitalsUpdateFrequency,
-    PropagatorType
+    VitalsUpdateFrequency
 } from './DdSdkReactNativeConfiguration';
 import { DdSdkReactNative } from './DdSdkReactNative';
 import { InternalLog } from './InternalLog';
@@ -19,7 +18,7 @@ import { TrackingConsent } from './TrackingConsent';
 import { DdTrace } from './foundation';
 import { DdLogs } from './logs/DdLogs';
 import { DdRum } from './rum/DdRum';
-import { RumActionType, ErrorSource } from './rum/types';
+import { RumActionType, ErrorSource, PropagatorType } from './rum/types';
 import { DatadogProvider } from './sdk/DatadogProvider/DatadogProvider';
 
 export {
