+++ packages/core/src/rum/instrumentation/resourceTracking/domain/__tests__/firstPartyHosts.test.ts
@@ -4,7 +4,7 @@
  * Copyright 2016-Present Datadog, Inc.
  */
 
-import { PropagatorType } from '../../../../../DdSdkReactNativeConfiguration';
+import { PropagatorType } from '../../../../types';
 import { firstPartyHostsRegexMapBuilder } from '../firstPartyHosts';
 
 describe('firstPartyHosts', () => {
