+++ packages/core/src/rum/instrumentation/resourceTracking/domain/distributedTracing.tsx
@@ -4,7 +4,7 @@
  * Copyright 2016-Present Datadog, Inc.
  */
 
-import type { PropagatorType } from '../../../../DdSdkReactNativeConfiguration';
+import type { PropagatorType } from '../../../types';
 
 import type { Hostname } from './firstPartyHosts';
 import { getPropagatorsForHost } from './firstPartyHosts';
